Expose the actual task object as part of the snapshot inputs operation details.
This pulled a thread and required reworking the build operations fixture to have the operations serialized from inside the build.
This allows more flexible serialization strategies to be used.
It also has the benefit of capturing the entire operation stream, not just the operations since the fixture init script ran.